requirements.txt
numpy~=1.24.4
keras~=2.10.0
tensorflow~=2.10.1
matplotlib~=3.7.5
pillow~=10.3.0